Elimination of some tests in src/common and addition of the test for the matching_engine.
